+++ Arbiter.App/Mappings/Server/ServerMessageMappingProvider.cs
@@ -58,6 +58,7 @@ public class ServerMessageMappingProvider : IInspectorMappingProvider
         RegisterServerShowNotepadMapping(registry);
         RegisterServerShowSpinnerMapping(registry);
         RegisterServerShowUserMapping(registry);
+        RegisterServerStatPointsMapping(registry);
         RegisterServerStatusEffectMapping(registry);
         RegisterServerSwitchPaneMapping(registry);
         RegisterServerSyncTicksMapping(registry);
@@ -831,6 +832,17 @@ public class ServerMessageMappingProvider : IInspectorMappingProvider
                 .Property(m => m.MonsterSprite, p => p.ToolTip("Sprite to display the user as a monster."))
                 .Property(m => m.MonsterUnknown)
                 .IsExpanded(m => m.HeadSprite == 0xFFFF);
+        });
+    }
+
+    private static void RegisterServerStatPointsMapping(InspectorMappingRegistry registry)
+    {
+        registry.Register<ServerStatPointsMessage>(b =>
+        {
+            b.Section("Stat Points")
+                .Property(m => m.FlashButtons,
+                    p => p.ToolTip("Whether the stat buttons should play their flash animation."))
+                .Property(m => m.StatPoints, p => p.ToolTip("Number of available stat points."));
         });
     }